Fixed the tests not working
