+++ src/components/InfinityMenu/infinityMenu.jsx
@@ -19,7 +19,7 @@ import Defiant from 'defiant.js';
 import Sticky from 'react-sticky-el';
 
 import SearchInput from './SearchInput';
-import ClassificationLink from './classificationLink';
+import ClassificationLink from './ClassificationLink';
 import EmptyTree from './EmptyTree';
 import Exporter from '../Exporter';
