+++ src/index.jsx
@@ -7,23 +7,15 @@ import 'react-app-polyfill/stable';
 import 'fast-text-encoding';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { createBrowserHistory } from 'history';
-import PiwikReactRouter from 'piwik-react-router';
 import * as Sentry from '@sentry/react';
 import { BrowserTracing } from '@sentry/tracing';
 import { registerLocale, setDefaultLocale } from 'react-datepicker';
 import fi from 'date-fns/locale/fi';
 
-import createStore from './store/createStore';
 import AppContainer from './containers/AppContainer';
 import config from './config';
 import routes from './routes';
-
-// Piwik Configuration
-const piwik = PiwikReactRouter({
-  url: config.PIWIK_URL,
-  siteId: config.PIWIK_ID,
-});
+import { store, history } from './init'
 
 // Register a locale for all datepickers in the application
 registerLocale('fi', fi);
@@ -36,15 +28,6 @@ Sentry.init({
   tracesSampler: 1.0,
 });
 
-// ========================================================
-// Store Instantiation
-// ========================================================
-const browserHistory = createBrowserHistory();
-const initialState = window.___INITIAL_STATE__;
-
-export const store = createStore(browserHistory, initialState);
-
-const history = piwik.connectToHistory(browserHistory);
 
 // // ========================================================
 // // Go!
